scale StoreTable
